Move EmulatorState enum
