+++ extension/src/extension.ts
@@ -1,5 +1,5 @@
 /* The extension */
-import { EmulatorState, EmulatorController } from './emulator/emulator-controller'
+import { EmulatorController } from './emulator/emulator-controller'
 import { CommandController } from './commands/command-controller'
 import { Account } from './emulator/account'
 import { UIController } from './ui/ui-controller'
@@ -7,6 +7,7 @@ import { ExtensionContext } from 'vscode'
 import { DEBUG_LOG } from './utils/debug'
 import { DependencyInstaller } from './dependency-installer/dependency-installer'
 import { Settings } from './settings/settings'
+import { EmulatorState } from './emulator/server/language-server'
 
 // The container for all data relevant to the extension.
 export class Extension {
